Fix layout hierarchy for sidebar display.
